refactor(admin): consolidate admin pages into role-based components
