+++ src/features/admin/components/shared/columns.module.css
@@ -8,5 +8,6 @@
 }
 
 .actionIcon {
-  size: 14;
+  width: var(--icon-size-md);
+  height: var(--icon-size-md);
 }
