--- src/features/admin/pages/DepartmentRequestPage/DepartmentRequestPage.module.css
@@ -1,55 +1,0 @@
-.loadingContainer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100vh;
-}
-
-.page {
-  display: grid;
-  grid-template-rows: auto 1fr auto;
-  width: 100%;
-  min-height: 100vh;
-  background-color: var(--color-bg);
-}
-
-.main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-lg);
-  width: 100%;
-  max-width: var(--container-max-width);
-  padding: var(--space-lg);
-}
-
-@media (min-width: 768px) {
-  .mainContainer {
-    padding: var(--space-2xl) var(--space-7xl);
-  }
-}
-
-.titleHeader {
-  width: 100%;
-  margin-bottom: var(--space-md);
-  font-size: var(--font-size-xl);
-  font-weight: var(--font-weight-bold);
-  color: var(--color-text);
-  text-align: center;
-}
-@media (min-width: 768px) {
-  .titleHeader {
-    font-size: var(--font-size-2xl);
-    text-align: left;
-  }
-}
-
-.tableSection {
-  width: 100%;
-}
