--- src/features/admin/pages/DepartmentResearchPage/DepartmentResearchPage.module.css
@@ -1,81 +1,0 @@
-.loadingContainer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100vh;
-}
-
-.page {
-  display: grid;
-  grid-template-rows: auto 1fr auto;
-  width: 100%;
-  min-height: 100dvh;
-}
-
-.main {
-  display: block;
-}
-
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-md);
-  max-width: var(--container-max-width);
-  padding: var(--space-lg);
-  margin: 0 auto;
-}
-
-@media (min-width: 768px) {
-  .mainContainer {
-    padding: var(--space-2xl) var(--space-7xl);
-  }
-}
-
-.headerSection {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-lg);
-}
-
-@media (min-width: 768px) {
-  .headerSection {
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-  }
-}
-
-.titleHeader {
-  font-size: var(--font-size-2xl);
-  font-weight: var(--font-weight-bold);
-}
-
-.createButton {
-  width: 100%;
-}
-@media (min-width: 768px) {
-  .createButton {
-    align-self: auto;
-    width: auto;
-  }
-}
-
-.tabsContainer {
-  display: flex;
-  gap: var(--space-lg);
-  margin: var(--space-lg) 0;
-}
-
-.tabButton {
-  display: flex;
-  flex: 1;
-  gap: var(--space-xs);
-  align-items: center;
-  justify-content: center;
-}
-
-.iconTab {
-  width: var(--icon-size-md);
-  height: var(--icon-size-md);
-}
